+++ plugins/documentation/readme-badges/main/kotlin/com/javiersc/gradle/plugins/readme/badges/generator/models/MavenRepo.kt
@@ -1,0 +1,7 @@
+package com.javiersc.gradle.plugins.readme.badges.generator.models
+
+/** Sonatype Maven repos */
+internal enum class MavenRepo {
+    MavenCentral,
+    Snapshot
+}